requirements.txt
django==6.0.4
djangorestframework==3.17.1
django-unfold==0.90.0
psycopg[binary]>=3.2,<3.3
gunicorn>=23.0,<24.0
google-genai>=1.0.0,<2.0.0
drf-spectacular>=0.27.0,<0.28.0
drf-spectacular-sidecar>=2024.7.1
